LWJGL vr support
Conflicts:
	jme3-vr/src/main/java/com/jme3/input/vr/openvr/OpenVRInput.java
	jme3-vr/src/main/java/com/jme3/input/vr/openvr/OpenVRViewManager.java

Conflicts:
	jme3-vr/src/main/java/com/jme3/input/vr/openvr/OpenVRInput.java
	jme3-vr/src/main/java/com/jme3/input/vr/openvr/OpenVRViewManager.java
