+++ jme3-vr/build.gradle
@@ -2,7 +2,7 @@ if (!hasProperty('mainClass')) {
     ext.mainClass = ''
 }
 
-def lwjglVersion = '3.1.3'
+def lwjglVersion = '3.2.0'
 
 sourceCompatibility = '1.8'
 
@@ -18,4 +18,9 @@ dependencies {
     // Native LibOVR/Oculus support
     compile "org.lwjgl:lwjgl-ovr:${lwjglVersion}"
     runtime "org.lwjgl:lwjgl-ovr:${lwjglVersion}:natives-windows"
+    
+    compile "org.lwjgl:lwjgl-openvr:${lwjglVersion}"
+    compile "org.lwjgl:lwjgl-openvr:${lwjglVersion}:natives-windows"
+    compile "org.lwjgl:lwjgl-openvr:${lwjglVersion}:natives-linux"
+    compile "org.lwjgl:lwjgl-openvr:${lwjglVersion}:natives-macos"
 }
