+++ jme3-vr/src/main/java/com/jme3/input/vr/openvr/OpenVRTrackedController.java
@@ -3,7 +3,6 @@ package com.jme3.input.vr.openvr;
 import com.jme3.app.VREnvironment;
 import com.jme3.input.vr.VRInputAPI;
 import com.jme3.input.vr.VRTrackedController;
-import com.jme3.input.vr.lwjgl.OpenVrLwjgl;
 import com.jme3.math.Matrix4f;
 import com.jme3.math.Quaternion;
 import com.jme3.math.Vector3f;
@@ -78,7 +77,7 @@ public class OpenVRTrackedController implements VRTrackedController{
 		
 		if (environment != null){
 			 if (hardware != null){
-					return ((OpenVrLwjgl)environment.getVRHardware()).getPoseMatrices()[controllerIndex];
+					return ((OpenVR)environment.getVRHardware()).getPoseMatrices()[controllerIndex];
 				} else {
 					throw new IllegalStateException("No underlying VR API.");
 				}
