+++ jme3-vr/src/main/java/test/TestInitHmd.java
@@ -32,7 +32,7 @@
 package test;
 
 import com.jme3.app.VREnvironment;
-import com.jme3.input.vr.lwjgl.OpenVrLwjgl;
+import com.jme3.input.vr.openvr.OpenVR;
 import com.jme3.system.AppSettings;
 
 /**
@@ -48,7 +48,7 @@ public class TestInitHmd {
     public static void testInitHmd(){
         VREnvironment environment = new VREnvironment(new AppSettings(true));
         environment.initialize();
-        OpenVrLwjgl openVr = (OpenVrLwjgl) environment.getVRHardware();
+        OpenVR openVr = (OpenVR) environment.getVRHardware();
         System.out.println(openVr.getName());
         
         openVr.updatePose();
